+++ flutter_embed_unity_2022_3_android/android/src/main/kotlin/com/learntoflutter/flutter_embed_unity_android/messaging/SendToUnity.kt
@@ -1,10 +1,13 @@
 package com.learntoflutter.flutter_embed_unity_android.messaging
 
+import com.learntoflutter.flutter_embed_unity_android.constants.FlutterEmbedConstants.Companion.logTag
 import com.learntoflutter.flutter_embed_unity_android.constants.FlutterEmbedConstants.Companion.methodNamePauseUnity
 import com.learntoflutter.flutter_embed_unity_android.constants.FlutterEmbedConstants.Companion.methodNameResumeUnity
 import com.learntoflutter.flutter_embed_unity_android.constants.FlutterEmbedConstants.Companion.methodNameSendToUnity
+import com.learntoflutter.flutter_embed_unity_android.constants.FlutterEmbedConstants.Companion.methodNameUnloadUnity
 import com.learntoflutter.flutter_embed_unity_android.unity.UnityPlayerSingleton
 import com.unity3d.player.UnityPlayer
+import io.flutter.Log
 import io.flutter.plugin.common.MethodCall
 import io.flutter.plugin.common.MethodChannel
 
@@ -24,6 +27,14 @@ class SendToUnity: MethodChannel.MethodCallHandler {
             }
             methodNameResumeUnity -> {
                 UnityPlayerSingleton.getInstance()?.resume()
+            }
+            methodNameUnloadUnity -> {
+                // On Android a true unload is not possible: UnityPlayer.destroy() would
+                // kill the whole app process (it was designed to own its own activity /
+                // process). So the best we can do is pause Unity to reduce resource usage.
+                Log.i(logTag, "unloadUnity requested. On Android Unity cannot be unloaded " +
+                        "without killing the app process, so pausing Unity instead.")
+                UnityPlayerSingleton.getInstance()?.pause()
             }
             else -> {
                 result.notImplemented()