+++ flutter_embed_unity_2022_3_ios/ios/flutter_embed_unity_2022_3_ios/Sources/flutter_embed_unity_2022_3_ios/Constants/FlutterEmbedConstants.swift
@@ -4,4 +4,6 @@ struct FlutterEmbedConstants {
     static let methodNameSendToFlutter = "sendToFlutter"
     static let methodNamePauseUnity = "pauseUnity"
     static let methodNameResumeUnity = "resumeUnity"
+    static let methodNameUnloadUnity = "unloadUnity"
+    static let creationParamUnloadOnDispose = "unloadOnDispose"
 }