+++ flutter_embed_unity_6000_0_android/android/src/main/kotlin/com/learntoflutter/flutter_embed_unity_android/constants/FlutterEmbedConstants.kt
@@ -11,5 +11,7 @@ class FlutterEmbedConstants {
         const val methodNameSendToFlutter = "sendToFlutter"
         const val methodNamePauseUnity = "pauseUnity"
         const val methodNameResumeUnity = "resumeUnity"
+        const val methodNameUnloadUnity = "unloadUnity"
+        const val creationParamUnloadOnDispose = "unloadOnDispose"
     }
 }