+++ flutter_embed_unity_platform_interface/lib/method_channel_flutter_embed_unity.dart
@@ -28,4 +28,11 @@ class MethodChannelFlutterEmbedUnity extends FlutterEmbedUnityPlatform {
       FlutterEmbedConstants.methodNameResumeUnity,
     );
   }
+
+  @override
+  void unloadUnity() {
+    _channel.invokeMethod(
+      FlutterEmbedConstants.methodNameUnloadUnity,
+    );
+  }
 }