+++ flutter_embed_unity_platform_interface/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 2.1.0
+
+* Added the `unloadUnity()` method to the platform interface, to unload Unity from memory rather than just pausing it.
+
+
 ## 2.0.0
 
 * Increased minimum Flutter version to 3.35 due to problems with Unity, Xcode 26 and Flutter JIT. See [#73](https://github.com/learntoflutter/flutter_embed_unity/issues/73)